LG-14609 Stop logging Threatmetrix response body in verify info results (#11294)
We use Threatmetrix for device profiling. When a user's device profiling transaction represents a review or reject status they need to undergo the fraud review process. Part of that process is looking at what was returned by Threatmetrix in the logged response body.

The Threatmetrix response body is incredibly large. It causes some issues with Cloudwatch parsing the JSON written to the log.

In 2ddece00b322a3a38c577cbc172e831700200dd1 we started logging the response body on its own event. We continued to log the Threatmetrix response body in its original place on the verify info proofing results event. This was done so we can validate that the response body is properly logged on the new event and works for our purposes.

This commit follows up that previous commit by removing the response body from the original event. We can merge this once we have verified that everything is working as expected with the new event.


[skip changelog]