LG-14716 Download and store Socure reason codes weekly (#11350)
When Socure responds to requests against its identity verification APIs it includes "reason codes" to describe why a transaction may have failed. These reason codes map back to descriptions that are available in the Reason Code API.

Socure has suggested to us that we maintain our own copy of their reason codes. This commit adds a background job that will create/update a copy of the Socure reason codes and descriptions in our database on a weekly basis.

The job fetches the list of codes from the API. It adds any new codes that do not exist in the database. It deactivates any codes that we have in the database that we do not observe in the response.

The job will log the results when it is done. This includes:

1. A list of reason codes that were added
2. A list of reason codes that were deleted
3. Any exceptions that may have occurred when adding the codes.

This change includes a `idv_socure_reason_code_download_enabled` config that controls whether this feature is active. It defaults to `false` since we are not currently utilizing Socure in production.

changelog: Internal, Socure, Download and store reason codes weekly