Log the user of a fraud rejected profile on automatic rejection (#11384)
If a user is in fraud review for over 30 days they are automatically rejected by the `FraudRejectionDailyJob`. Prior to this commit the analytics service was initialized with an anonymous user. As a result the event was not logged for the fraud review rejected user. This commit fixes that issue so these events will appear associated with the rejected user.

[skip changelog]