Update DIVR to do fraud stats for SPs (#11311)
* Update DIVR to do fraud stats for SPs

Several fraud-related analytics events do not have SP information associated with them.
Filtering by SP for ther DIVR will therefore exclude those events.

This PR updates the DIVR to pull all fraud-related events for the given period for comparison.

[skip changelog]

* Try to be better

* Update passed_fraud_review_users logic

When filtering by SP, only count users we know can be attributed to the SP, either because:

1. They have a final resolution event in the dataset attributed to the SP or
2. The profile review passed event includes a matching service_provider

For a while, we will have some fraud events augmented with service_provider and some without.

* Apply SP logic to did_not_pass_fraud_review_users

* Guard against nil issuers array

* Address linting issues

* Pull all GPO verification events in the period

We need these events returned so that we can attribute fraud events that are lacking SP data.

* Pull all IPP events in time period

Similar to GPO, we need the IPP events even if they are fraud review pending so that we can attribute fraud-related events that are lacking SP data.

* Add a user who bounced on the welcome screen

* Remove unused constant

* Remove unused SP event tracking

* Clean up sp_key a little

* Ensure we filter in service_provider

service_provider is normalized using `coalesce` whereas properties.service_provider is not

* Clean up diff a little hopefully

* Refine how we normalize fraud_review_pending

- use coalesce() + OR to ensure we get rid of all null values

* Remove fraud_pending_reason consideration

fraud_pending_reason is sourced from the Profile for Idv: final resolution and may have a value set even if the profile is not currently in the fraud review queue (e.g. for GPO).

* Revert "Remove fraud_pending_reason consideration"

Actually, we want to consider this--the only way to identity gpo / fraud pending is by looking at fraud pending reason

This reverts commit 9dc96630f210964f42a4f0b5f0f7e49625cf60c3.

* Add a comment about fraud_review_pending usage

* delint

* Update lib/reporting/identity_verification_report.rb

Co-authored-by: Matt Wagner <matt.wagner@gsa.gov>

* Handle case where multiple issuers are specified

First build a list of all unique users for the specified issuers, then intersect that list with the list of users who passed fraud review. This way we are allowing for users that interacted with _any_ of the issuers rather than _all_ of them.

* Learn how each_with_object works

It's not like reduce, you have to modify the accumulator

* Add test for multiple issuer usecase

Test that, if results are returned for multiple issuers, we correctly count successfully passed users

---------

Co-authored-by: Jonathan Hooper <jonathan.hooper@gsa.gov>
Co-authored-by: Matt Wagner <matt.wagner@gsa.gov>