LG-14366 Store dynamically-generated proofing components on Profile
Proofing components are intended to represent the tooling that was used to create a profile. Traditionally these have been generated by tracking proofing components in a database table associated with the user. There were a number of problems with this approach which 5021e2ef8237ecc6ff38f6338371e89fb8dcf06d introduced a `Idv::ProofingComponents` service to address. This service generates proofing components from the session instead of from the database table.

This commit uses the `Idv::ProofingComponents` service to generate the proofing components stored on the profile. This will allow us to drop the proofing components table in a future change.

[skip changelog]