Fix a typo in the `ThreatMetrixPlugin`
It looks like in 1767dd586ed9fd1cab3662a90fece1e992d679a3 we introduced a typo where `request_ip` was spelled `request_id`. This resulted in us send a null `request_ip` to Threatmetrix.

This commit fixes the typo and modifies a spec to make sure the expected arguments are sent to the proofer.

[skip changelog]