LG-15125: Wait for Socure result - hybrid flow
As in #11500 for the standard flow, we need to wait for the Socure
result to be received before we can determine where to send the user.

Currently, if the wait times out, we show a placeholder plain text of
"Technical difficulties!!!" to be replaced with the appropriate "Try
again?" page in a following ticket.

Note there is some identical code in the standard vs hybrid controllers
that might be refactored into a concern in the future, but until all of
the paths are complete, I thought it best to live with the duplication
for now.

[skip changelog]