Add name suffix to TrueID response fixture
LexisNexis informed us that the name suffix should be available in a TrueID response if the name suffix is present on the document. This commit adds the name suffix to the response fixture so that we have a fixture to represent this case.

We plan to read the name suffix and check it against the DLDV service in a future change. This commit prepares us for that.

[skip changelog]