Add new fields to `Pii::StateId`
We are planning to validate additional data that appears on the state ID with the issuing source via DLDV. In order to facilitate this new feature this change adds fields for the new attributes to the `StateId` struct. This struct is used to represent the State ID data when it is in the IdV session.

[skip changelog]