LG-14985 Read additional document data from TrueID when configured to do so (#11559)
We are working to send additional data to AAMVA when it is present on documents. This commit lays groundwork for this by reading those new attributes when configured to do so.

We are waiting on security approval before we start writing these new attributes into the Login.gov session. For that reason the reading of these attributes is feature flagged and disabled by default.

changelog: Internal, Document Authentication, Read additional document data from TrueID when configured to do so

* cleanup