LG-13763: Stop writing to the proofing_components table (#11564)
* Stop writing to the proofing_components table

Just stop.

These values are now generated dynamically as we need them, then serialized to the Profile.

[skip changelog]

* Update LinkSentController spec

* Update UspsLocationsController + spec

I moved idv_session to be public on the controller (matching the behavior of other controllers that use IdvSessionConcern).

Then updated tests to use dynamically-generated proofing components. This requires reloading the user object. I created https://cm-jira.usa.gov/browse/LG-15167 to address computing the proofing component value without using the User object

* Move proofing_component check to end of end-to-end spec

We don't have access to idv_session here so we have to wait until the proofing component has been written to the database so we can access it.

* Remove proofing components from pw reset disavowal form spec

We were clearing out the _user's_ proofing component here, but not doing anything with the profile.

* Remove references to ProofingComponent from specs

* Don't check proofing_component in ResolutionProofingJob

* Remove ProofingComponent model