LG-14988 Send additional attributes to AAMVA (#11565)
We are planning on sending additional attributes that we read off of documents to AAMVA. This commit modifies the AAMVA client to do so. The new attributes are:

- Middle name
- Name suffix
- Sex
- Height
- Weight
- Eye color

These attributes are only sent if they are available in the applicant. For most of the attributes we do not currently read them from the document so they will not be sent until we enable the feature to do so. The exception to this is middle name. For that reason a feature flag controls whether we send middle name.

We will log whether these attributes are sent and whether they are validated. None of these attributes are "required attributes" so a user will still pass if they are absent or if they do not match.

changelog: Internal, AAMVA DLDV, Send additional attributes to AAMVA