LG-15177 Add a CTA to return to an SP after entering a verify-by-mail code
In #11591 we added a `post_idv_follow_up_url` that allows users to return to a service provider after finishing verification out-of-band.

This commit puts this new URL to work in the verify-by-mail code. After this commit user's who enter a verify-by-mail code without an SP in their session will see a new SP follow-up screen. This screen includes a CTA that prompts the user to connect their account to the service provider they started proofing with. This link leads to the new post-IdV follow-up URL.

changelog: User-Facing Improvements, Verify-by-mail, A CTA was added to prompt users to return to the service provider after verify-by-mail