LG-15121 Add a `post_idv_followup_url` configuration to service providers (#11591)
We have been working through issues with a service provider who has observed that users do not return to their SP after entering their mailed code or passing fraud review. We identified an opportunity to provide users with a direct link back to the SP to re-initiate their session when they complete out-of-band proofing.

This commit adds a `post_idv_followup_url` to the SP configuration for this purpose. This URL is used in the `account_verified` email if it is present. In future changes we plan to add a CTA after the user enters a mailed code and to the account page if the user has not connected their account.

This change should not impact SPs that do not have a `post_idv_followup_url` configured.

[skip changelog]