LG-15237 Allow the SP return URL resolver to interpolate locale values into URLs
In #11591 we added a `post_idv_follow_up` URL in response to issues we were having getting users for a particular partner to return to their SP. This partner expressed concerns about have the appropriate locale for these users since they are returning to the partner without initiating a session at the partner.

The original implementation for the change in #11591 included a locale param on the `post_idv_follow_up` URL. After some discussion with other engineers we identified that this pattern of hardcoding a param onto a single service providers was undesirable. This commit introduces the ability to include templateable values in these URLs and introduces the locale as one of those values.

changelog: Internal, SP Configuration, Add the ability to specify a locale in a redirect URL