Modify the step-indicator page such that the "Re-enter password" step is not marked complete
We received feedback from a partner that users may think that they are done with the identity verification workflow on the personal key page because all of the steps on the step indicator are marked complete. This commit modifies the step indicator on the personal key page so the "Re-enter password" step is not complete on that page.

Since there is not longer a case where all step indicator steps are complete this change also removes the `complete` status from the `StepIndicatorComponent`.

[skip changelog]