LG-15104 Skip enrollment checks when profile in password_reset
Since we are moving towards updating pending enrollments to be captured
by the password reset flow we need to update the job to not cancel
pending enrollments that have profiles with deactivation reason of
password_reset.

changelog: Internal, In-person Proofing, Update the get usps proofing results job to skip pending enrollments with profiles that have a deactivation reason of password_reset.