LG-15135 | Translations interpolate SP name
Due to a misunderstanding with DOS, the translations had the
literal string "service provider". Fixed!

Conditionally hides the display if there is no SP (which shouldn't
happen in prod) because otherwise the message is nonsense and has
no clickable link text.

changelog: User-Facing Improvements, Account screen, Show SP name in return-to-SP CTA