LG-15396 Log the initiating SP on verify-by-mail code entry
When a user enters a code during GPO there usually is not a service provider present in the SP session. This is because the letter we send instructs users to go directly to Login.gov to enter the code. As a result it is difficult to collect verification events for service providers in the out-of-band case since we cannot group events by `properties.service_provider`.

This commit adds the initiating service provider to the logged analytics event when a user enters a GPO code. The issuer is already logged on in-person and fraud review so this change allows us to group all out-of-band verification events by service provider.

changelog: Internal, Analytics, Initiating SP issuer is logged on verify-by-mail code entry