Drop proofing_components table
Since #11564 we've stopped writing to this table, so now is the time to drop it.

This table was previously used as a temporary datastore for proofing component values as a user moves through IdV. We now dynamically derive those values from the Idv::Session, so this table is not needed any longer.

[skip changelog]