Account for a space in the height returned by TrueID (#11737)
The regex we used to read the height from a TrueID response did not account for a space that may appear between feet and inches. As a result height values like `5' 11"` were not read. This commit updates the regex to fix that.

changelog: Internal, TrueID, Regex to parse the height from documents was adjusted