LG-15391 Allow SPs to request the UI locale
We have a service provider who would like to be able to receive the user's UI locale when they receive attributes after sign in. This partner has a `post_idv_follow_up_url` that is configured to redirect the user to Login.gov with a request for verified attributes. Since the user does not see a UI on their site and likely does not have an established session there they want to know what locale the user is using on Login.gov to provide a consistent experience.

This commit enables partners to request to UI locale using the `ui_locale` OIDC scope. The locale is stored in the user table and updated whenever the user changes their locale. This ensures the latest value will be returned to service providers.

changelog: Upcoming features, Requestable attributes, A claim that will allows SPs to request the user's UI locale was added