LG-15375: Ensure Socure user can opt in to IPP
We found a bug where if a user was bucketed into Socure for doc auth,
when they tried to opt-in to IPP from the How To Verify page, they would
be redirected to Socure. This fixes it.

changelog: Bug Fixes, Identity Verification, Allow users designated for Socure to opt-in to In Person Proofing