LG-15392 Remove the locale param from the post-IdV follow-up URL
We added the locale param to the post-IdV follow-up URL as a short term solution for a partner who needed to get the user's locale when they clicked that link. In #11756 we made a change that allows SPs to request the locale via the OIDC claims or SAML attributes. This should resolve the issue for that partner.

This commit removes support for including the locale as a param on the post-IdV follow-up URL. It should not be merged until:

1. Our partner has configured their application to request the locale in the OIDC params
2. We have removed the templated locale value from their configured post-IdV follow-up URL

[skip changelog]