LG-15394 Add tooling for an SP proofing events by UUID report
We have an agency partner who requested a specialized report the involves a table with columns representing proofing events and rows representing users for the partner where the values in each cell correspond to whether that user encountered that event in the given time period. We have been manually servicing this request for a little while. This commit adds tooling to generate this report automatically so it does not require manual processing.

This commit does not add the report to the job schedule with the intention being to test it first before scheduling it to run automatically.

[skip changelog]