LG-15395 Add specialized SP IdV Dropoff Report
We have a partner who requested a specialized IdV drop-off report for their service providers. This report needs to run every Monday and collect weekly data from a specified date up through the previous week.

We have been running this report manually for a few weeks now. This commit moves the code the generates this report into the IdP so it can be run automatically.

[skip changelog]