LG-14522 Enable password reset for pending in-person profiles (#11762)
* LG-14522 Enable password reset for pending in-person profiles

changelog: User-facing Improvement, In-person Proofing, Allow users to perform a password reset on pending in-person profiles without losing progress if the user supplies their personal key.

* Add feature flag to enable/disable pending in person password reset

* Remove activate_after_password_reset method from profile

* Fix pending_profile_policy for pending in person when facial match is required

Reproduce by creating an IPP enrollment arriving at the ready to verify
page and logout. Then log back in you will land on the welcome page.
This is incorrect because the user has a pending in person enrollment.

* Prioritize the pending profile over the active profile

The personal key only works for the most recent profile that had a
personal key generated

* Remove binding.pry

* Update the get usps proofing results job

Allow enrollments with password reset profiles to be updated when the
enrollment is not passed/failed. So that enrollments can expire when
still in password reset.

* ensure pending_profile_invalidated is correct in analytics

---------

Co-authored-by: Eileen McFarland <80347702+eileen-nava@users.noreply.github.com>
Co-authored-by: gina-yamada <gina.yamada@gsa.gov>