LG-15643 Pre-check phone numbers when the phone number is present in the applicant
We are working on a feature where the user's phone number can be verified in the verify-info check. This feature is called "Phone Finder Pre-check".

Phone-Finder Pre-check will make a best effort to find a phone number on the user's account (either from their MFA methods or hybrid flow) and validated that phone number when the user submits their info. If the number matches the user will be sent straight to OTP confirmation for that phone number.

As part of this implementation we will need the resolution proofing job to send the phone number to phone finder if a phone number is found for the user at the verify info step. This commit implements that by have the resolution proofing job communicate with phone finder and store a result if a phone number is provided in the applicant.

[skip changelog]