LG-15586 Remove the "Entered the wrong number" message from IdV OTP verification controller
As part of the phone-finder pre-check work the user can arrive on the OTP verification screen without entering a number. In this case the "Entered the wrong number" message on the screen does not make sense.

This commit removes the message and leaves the link that allows a user to try a different number.

[skip changelog]