LG-12326 Update the phone warning screen message
As part of the phone-finder pre-check work the user can arrive on the phone warning screen without entering a number. In this case the "You entered (XXX) XXX-XXXX" message on the screen does not make sense.

This commit changes the message to say "We couldn't find a record of you using this number: (XXX) XXX-XXXX".

[skip changelog]