LG-15573 Make the `--depth` argument required for IG data pull script
We often get requests for data that involve searching for accounts that share devices. These kinds of requests occur more often than not. Currently we default to not searching for shared devices. This has led to mistakes where we did not provide shared account data.

This commit attempts to address the issue by making the `--depth` argument required. This will require investigators to explicitly opt-in or out of shared device searches.

[skip changelog]