LG-15573 Make the `--depth` argument required for IG data pull script (#11960)
We often get requests for data that involve searching for accounts that share devices. These kinds of requests occur more often than not. Currently we default to not searching for shared devices. This has led to mistakes where we did not provide shared account data.

This commit attempts to address the issue by making the `--depth` argument required. This will require investigators to explicitly opt-in or out of shared device searches.

The `exit 1` in `lib/script_base.rb` was causing some rather opaque failures in CI as it sets the status code to 1 and caused some tests to end early rather without reporting whether they failed. The unhelpful message at the end of a CI run was Stopped processing SimpleCov as a previous error not related to SimpleCov has been detected. This only became an issue because the PR added a case where we raise an exception and it resulted in a non-zero status code.

[skip changelog]

Co-authored-by: Mitchell Henke <mitchell.henke@gsa.gov>