Allow approving pending IPP enrollments in int
Currently, when testing the IPP flow in the sandbox (int), one must wait 1 to 2
hours before the verification is complete. The sandbox is the environment
that all our partners use to test their apps, and that the Partnerships team
uses to test, and for live demos for partners. Having to wait 1 to 2 hours is a huge blocker.

A few months ago, a feature was built that allows approving pending IPP
enrollments immediately via the /test/ipp route. This feature was originally
built for local development only. This PR seeks to make this feature available
in our `int` host as well.

In the process, I made some improvements to the existing feature, such as:

- Add controller and feature tests
- Move the logic that determines whether or not this feature is available to `FeatureManagement`
- Make sure the user is fully authenticated
- In `int`, only display pending enrollments that belong to the current user
- Make queries more robust and performant with `find_by` and avoiding N+1 queries
- In the view, use `size` instead of `count`, since `count` can lead to some issues in some cases (although unlikely in this situation). See this true story: https://www.moncefbelyamani.com/the-6-characters-that-could-bring-down-your-rails-app/