Use the session locale on OIDC requests with no locale param
When partners initiate an authentication they can include a locale in the request which we honor when we render the next page. However, many partners do not do this and we assume the default locale of English.

If the user was previously on Login.gov they may have a locale in the session. This commit changes the IdP to honor that locale if no locale is present in the OIDC authorization request. This fixes the following weird behaviors:

_The step up flow_:

1. The user authenticates for auth-only and selects a non-English locale for that
2. The user is reidrected to the SP
3. The SP determines the user needs to identity proof and redirects back to Login.gov
4. The user's language choice is not longer respected

_The post-IdV follow up flow_:

1. The user enters a GPO code. They elect to do so with a non-English locale
2. They are presented the post-IdV follow up button on the next screen. They click the button to proceed to the service provider
3. The service provider recognizes it needs attributes from the user. It makes a redirect back to the IdP for the OIDC authorization endpoint
4. The user sees the screen to share proofed attributes. The language preference is no longer respected on this screen

This change causes the languge preference to be respected in both of these cases.

[skip changelog]