Use separate feature flag for logging duplicate profiles
changelog: Internal, One Verified Account, Use separate feature flag for logging duplicate profiles

Separating the feature flags so that we can turn on the recording of
duplicates without having to also have the user confirm the duplicate
accounts. This is intended to be a short-lived feature flag; we will
remove it as soon as we can confirm that it works as intended.