LG-16288: Guard against nil enrollment (#12288)
* LG-16288: Guard against nil enrollment

We have seen some errors where the enrollment passed to the IPP
verify_info_presenter is nil. This guards against that error, though it
is currently unclear how the enrollment could be nil here.

This may result in kicking the error further down the stack, but that
may give a clue as to where/how this is happening.

changelog: Bug Fixes, Passports, User prevented from advancing to steps before they are allowed.

* Add preconditions to existing IPP controllers to ensure that
the establishing enrollment exists.

Co-authored-by: Shane Chesnutt <shane.chesnutt@gsa.gov>

* specs

---------

Co-authored-by: Shane Chesnutt <shane.chesnutt@gsa.gov>