LG-16672 Fix rate limit in hybrid flow
When the user would submit successfully on the last attempt of doc auth
the application would rate limit the user. This was due to a race
condition between the javascript poller on the link sent page, and the
doc auth rate limit incrementing and processing time of the images. If
the image processing took longer than the polling time. The user would
be rate limited because the rate limit value was already incremented
when the value was polled.

changelog: Bug Fixes, Hybrid Doc Auth, Fix rate limiting bug for a successful last submission in IdV hybrid flow during doc auth.