Handle SAML request errors from saml_idp gem
Before, the saml_idp gem would return a 403 response if it found
specific errors with the SAML request, such as a missing issuer, or a
missing AuthnRequest. The problem is that this did not provide any
helpful information to partners while testing their integration.

To improve this, we updated the `saml_idp` gem to populate the
`saml_request` object with errors and not do anything else. This allows
the IdP to consume those errors and handle them appropriately, by
rendering them in the existing `/saml/auth/error` template.

This allows partners to quickly understand what part of their SAML
Request is  invalid, and makes it easier for them to fix the problem.
If they don't  know how to fix it themselves, they can let us know the
error message they see, which makes it easier and faster for us to help
our partners.

changelog: User-Facing Improvements, SAML Request Handling, Display SAML request errors so partners can understand what they did wrong