LG-17019: Update proofing components on AAMVA @ doc_auth
When a state ID is successfully validated, the source check in the
proofing components is updated with proofing source (e.g., AAMVA).

This is necessary when we check AAMVA at the DocAuth step.

If the source check is set by the time we get to the VerifyInfo step,
we skip the state id check in the progressive proofer.

changelog: Internal, Identity Verification, Record the state id check in the session to prevent redundant checks.

Co-authored-by: Alex Bradley <alexander.bradley@gsa.gov>
Co-authored-by: Shane Chesnutt <shane.chesnutt@gsa.gov>